Feat/web maintenance (#16)
* rewriting maintenance tool

* progress for now

* functioning gui

* only thing missing is reading the settings file and using the settings

* finished implementing usage of settings

* fixed script

* dummy implementation (that works ish) implemented for the last two buttons

* implemented comments

* whitespace n shit

+++ maintenance/rebuild.sh
@@ -2,7 +2,7 @@
 
 set -e
 if ! [ -f /etc/nixos/flake.nix ]; then
-    git clone https://github.com/f-klubben/Strandvejen -b final /etc/nixos
+    git clone https://github.com/f-klubben/Strandvejen /etc/nixos
     cd /etc/nixos
 else
     cd /etc/nixos
@@ -10,4 +10,7 @@ else
 fi
 nix flake update
 nixos-rebuild switch
-reboot
+
+if [[ $(jq -r ".should_restart" $MAINTENANCE_FILE) == "true" ]]; then
+    reboot
+fi
